+++ android-test/app/src/test/java/co/elastic/apm/android/test/centralconfigs/logs/OpenTelemetryLogRecordsTest.java
@@ -1,4 +1,4 @@
-package co.elastic.apm.android.test.opentelemetry.logs;
+package co.elastic.apm.android.test.centralconfigs.logs;
 
 import static org.mockito.Mockito.doReturn;
 
@@ -8,7 +8,7 @@ import org.robolectric.annotation.Config;
 import co.elastic.apm.android.sdk.internal.configuration.Configurations;
 import co.elastic.apm.android.sdk.internal.configuration.impl.AllInstrumentationConfiguration;
 import co.elastic.apm.android.sdk.logs.ElasticLoggers;
-import co.elastic.apm.android.test.opentelemetry.common.AppUseCases;
+import co.elastic.apm.android.test.centralconfigs.common.AppUseCases;
 import co.elastic.apm.android.test.testutils.base.BaseRobolectricTest;
 import io.opentelemetry.api.logs.Logger;
